productproject/requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.2.5
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
flower==2.0.1
humanize==4.8.0
kombu==5.3.2
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==4.6.0
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6